fixing problems with typescript
