+++ Bookcatalog.WebReact/package.json
@@ -7,20 +7,21 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.18.64",
+    "@types/react": "^18.2.38",
+    "@types/react-dom": "^18.2.17",
     "apollo-boost": "^0.1.22",
     "axios": "^0.18.0",
     "bootstrap": "^5.3.2",
-    "graphql": "^14.0.2",
     "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.9.5",
+    "web-vitals": "^2.1.4",
     "react-apollo": "^2.3.2",
     "react-bootstrap": "^2.9.1",
-    "react-dom": "^18.2.0",
-    "react-redux": "^6.0.0",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "5.0.1",
-    "redux": "^4.0.1",
-    "sass": "^1.69.5",
-    "web-vitals": "^2.1.4"
+    "sass": "^1.69.5"
   },
   "scripts": {
     "start": "react-scripts start",
